20260404: Add revenue type radio toggle, product autocomplete in line items, price lookup endpoint, engagement autocomplete fix, location autocomplete, autoPopulate ProductPricePlanId fix, purchase order line items and receipts

+++ views/purchase_order/list/page.go
@@ -131,7 +131,7 @@ func buildTableRows(orders []*purchaseorderpb.PurchaseOrder, l centymo.Expenditu
 		poNumber := po.GetPoNumber()
 		currency := po.GetCurrency()
 		recordStatus := po.GetStatus()
-		totalAmount := centymo.FormatCentavoAmount(float64(po.GetTotalAmount()), currency)
+		totalAmount := centymo.FormatCentavoAmount(po.GetTotalAmount(), currency)
 		orderDate := po.GetOrderDateString()
 		expectedDelivery := po.GetExpectedDeliveryDateString()
